LayoutTests/inspector/elements/event-listeners-about-blank.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function setupEventListeners()
{
    function f() {}
    var frame = document.getElementById("myframe");
    var body = frame.contentDocument.body;
    body.addEventListener("click", f, true);
    var div = frame.contentDocument.createElement("div");
    div.id = "div-in-iframe";
    div.addEventListener("hover", f, true);
    body.appendChild(div);
}

function test()
{
    WebInspector.settings.eventListenersFilter.set("all");
    InspectorTest.evaluateInPage("setupEventListeners()", step1);

    function step1()
    {
        InspectorTest.selectNodeWithId("div-in-iframe", step2);
    }

    function step2()
    {
        InspectorTest.expandAndDumpSelectedElementEventListeners(InspectorTest.completeTest);
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests event listeners output in the Elements sidebar panel when the listeners are added on an element in about:blank page.
</p>

<iframe id="myframe"></iframe>

</body>
</html>
