LayoutTests/inspector/elements/event-listeners-about-blank-expected.txt

Tests event listeners output in the Elements sidebar panel when the listeners are added on an element in about:blank page.



======== click ========
[expanded] body event-listeners-about-blank.html:9
    isAttribute: false
    lineNumber: 9
    listenerBody: function f() {}
    node: <body>
    sourceName: [clipped-for-test]/LayoutTests/inspector/elements/event-listeners-about-blank.html
    type: click
    useCapture: true

======== hover ========
[expanded] div#div-in-iframe event-listeners-about-blank.html:9
    isAttribute: false
    lineNumber: 9
    listenerBody: function f() {}
    node: <div>
    sourceName: [clipped-for-test]/LayoutTests/inspector/elements/event-listeners-about-blank.html
    type: hover
    useCapture: true
